Moved tests out of the files themselves to match the style of the emulation core.
