+++ src/tests.rs
@@ -1,1 +1,2 @@
 pub mod emulation_core;
+pub mod parser;
