+++ src/tests/parser.rs
@@ -1,0 +1,2 @@
+pub mod preprocessing;
+pub mod operand_reading;
